Move logrocket initialization to its own plugin
